Migrate default config values into field decl.
Fixes #20

The tricky ones are the ones where the default value is true. In order
to make it work appropriately, the best approach was to check for the
Config value being not equal to a ConfigObject, since by virtue of
checking that Config value it becomes a ConfigObject if it is not set.